+++ mmh3tools/__init__.py
@@ -10,6 +10,7 @@ from .nodes_music_analysis import MMH3MusicAnalysis
 from .nodes_musicscene import MMH3MusicScenePlanPrompt
 from .nodes_stylepack import MMH3LoadSkill
 from .nodes_controlnet import MMH3CondSetApplyControl
+from .nodes_embedselect import MMH3EmbeddingSelect
 from .nodes_encode import MMH3StreamingEncode
 from .nodes_lint import MMH3PromptLint
 from .nodes_music import (MMH3LyricsSectionize, MMH3MusicCaptionSplit,
@@ -78,6 +79,7 @@ NODES = [
     MMH3ContextWindows,
     # MMH3Tools/calculators
     MMH3ChunkSchedule,
+    MMH3EmbeddingSelect,
     MMH3ImageList,
     MMH3CondSetApplyControl,
     WhisperAlignmentToText,